+++ src/Domain/Settings/Actions/TestPostmarkConnectionAction.php
@@ -6,30 +6,17 @@ use Domain\Settings\Mail\TestMail;
 use Symfony\Component\Mailer\Exception\LogicException;
 use Symfony\Component\Mailer\Exception\TransportException;
 
-class TestMailgunConnectionAction
+class TestPostmarkConnectionAction
 {
-    /**
-     * Throw an Exception if connection isn't successful.
-     *
-     * @return never
-     */
     public function __invoke(array $credentials)
     {
         try {
             // Set temporary mail connection
             config([
-                'mail'     => [
-                    'driver'       => 'mailgun',
-                    'from.address' => $credentials['sender'],
-                    'from.name'    => $credentials['sender'],
-                ],
-                'services' => [
-                    'mailgun' => [
-                        'domain'   => $credentials['domain'],
-                        'secret'   => $credentials['secret'],
-                        'endpoint' => $credentials['endpoint'],
-                    ],
-                ],
+                'mail.driver'             => 'postmark',
+                'mail.from.address'       => $credentials['sender'],
+                'mail.from.name'          => $credentials['sender'],
+                'services.postmark.token' => $credentials['token'],
             ]);
 
             // Send test email