+++ src/Domain/Settings/Controllers/StoreEmailCredentialsController.php
@@ -5,11 +5,13 @@ use Artisan;
 use Illuminate\Http\JsonResponse;
 use Domain\Settings\Actions\TestSMTPConnectionAction;
 use Domain\Settings\Actions\TestMailgunConnectionAction;
+use Domain\Settings\Actions\TestPostmarkConnectionAction;
 use Domain\Settings\Requests\StoreEmailCredentialsRequest;
 
 class StoreEmailCredentialsController
 {
     public function __construct(
+        private TestPostmarkConnectionAction $testPostmarkConnection,
         private TestMailgunConnectionAction $testMailgunConnection,
         private TestSMTPConnectionAction $testSMTPConnection,
     ) {
@@ -38,16 +40,23 @@ class StoreEmailCredentialsController
                     'domain'   => $request->input('mailgun.domain'),
                     'secret'   => $request->input('mailgun.secret'),
                     'endpoint' => $request->input('mailgun.endpoint'),
+                    'sender'   => $request->input('mailgun.sender'),
+                ]),
+                'postmark' => ($this->testPostmarkConnection)([
+                    'token'  => $request->input('postmark.token'),
+                    'sender' => $request->input('postmark.sender'),
                 ]),
             };
 
             $mail = [
                 'log'      => [
-                    'MAIL_DRIVER' => 'log',
+                    'MAIL_DRIVER'       => 'log',
                 ],
                 'postmark' => [
-                    'MAIL_DRIVER'    => 'postmark',
-                    'POSTMARK_TOKEN' => $request->input('postmark.token'),
+                    'MAIL_DRIVER'       => 'postmark',
+                    'POSTMARK_TOKEN'    => $request->input('postmark.token'),
+                    'MAIL_FROM_ADDRESS' => $request->input('postmark.sender'),
+                    'MAIL_FROM_NAME'    => $request->input('postmark.sender'),
                 ],
                 'smtp'     => [
                     'MAIL_DRIVER'       => 'smtp',
@@ -56,8 +65,8 @@ class StoreEmailCredentialsController
                     'MAIL_USERNAME'     => $request->input('smtp.username'),
                     'MAIL_PASSWORD'     => $request->input('smtp.password'),
                     'MAIL_ENCRYPTION'   => $request->input('smtp.encryption') ?? '',
-                    'MAIL_FROM_ADDRESS' => $request->input('smtp.email') ?? '"${MAIL_USERNAME}"',
-                    'MAIL_FROM_NAME'    => $request->input('smtp.email') ?? '"${MAIL_USERNAME}"',
+                    'MAIL_FROM_ADDRESS' => $request->input('smtp.email') ?? $request->input('smtp.username'),
+                    'MAIL_FROM_NAME'    => $request->input('smtp.email') ?? $request->input('smtp.username'),
                 ],
                 'ses'      => [
                     'MAIL_DRIVER'           => 'ses',
@@ -67,10 +76,12 @@ class StoreEmailCredentialsController
                     'AWS_SESSION_TOKEN'     => $request->input('ses.session_token'),
                 ],
                 'mailgun'  => [
-                    'MAIL_DRIVER'      => 'mailgun',
-                    'MAILGUN_DOMAIN'   => $request->input('mailgun.domain'),
-                    'MAILGUN_SECRET'   => $request->input('mailgun.secret'),
-                    'MAILGUN_ENDPOINT' => $request->input('mailgun.endpoint'),
+                    'MAIL_DRIVER'       => 'mailgun',
+                    'MAILGUN_DOMAIN'    => $request->input('mailgun.domain'),
+                    'MAILGUN_SECRET'    => $request->input('mailgun.secret'),
+                    'MAILGUN_ENDPOINT'  => $request->input('mailgun.endpoint'),
+                    'MAIL_FROM_ADDRESS' => $request->input('mailgun.sender'),
+                    'MAIL_FROM_NAME'    => $request->input('mailgun.sender'),
                 ],
             ];